# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Multi-Store Inventory Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /production-batches:
    get:
      operationId: listProductionBatches
      tags: [production]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [in_progress, completed, cancelled]
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of production batches
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionBatchList"
    post:
      operationId: createProductionBatch
      tags: [production]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductionBatchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionBatch"
  /production-batches/{id}:
    get:
      operationId: getProductionBatch
      tags: [production]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Production batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionBatch"
  /production-batches/{id}/complete:
    post:
      operationId: completeProductionBatch
      tags: [production]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteBatchInput"
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionBatch"

  # SALES
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      parameters:
        - name: saleId
          in: query
          schema:
            type: integer
        - name: customerId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentList"
    post:
      operationId: createPayment
      tags: [payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # INVENTORY
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      parameters:
        - name: storeId
          in: query
          schema:
            type: integer
        - name: productId
          in: query
          schema:
            type: integer
        - name: lowStock
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Inventory stock levels
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryList"
  /inventory/movements:
    get:
      operationId: listInventoryMovements
      tags: [inventory]
      parameters:
        - name: storeId
          in: query
          schema:
            type: integer
        - name: productId
          in: query
          schema:
            type: integer
        - name: movementType
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Inventory movement log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryMovementList"
  /inventory/adjustments:
    post:
      operationId: createInventoryAdjustment
      tags: [inventory]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryAdjustmentInput"
      responses:
        "201":
          description: Adjustment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryMovement"

  # DASHBOARD
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationList"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    ProductionInput:
      type: object
      required: [productId, quantity, unit]
      properties:
        productId:
          type: integer
        productName:
          type: ["string", "null"]
        quantity:
          type: number
        unit:
          type: string
    ProductionBatch:
      type: object
      required: [id, batchNumber, type, status, stageFromStoreId, stageToStoreId, plannedOutputQty, createdAt]
      properties:
        id:
          type: integer
        batchNumber:
          type: string
        type:
          type: string
          enum: [raw_to_semi, semi_to_finished]
        status:
          type: string
          enum: [in_progress, completed, cancelled]
        stageFromStoreId:
          type: integer
        stageToStoreId:
          type: integer
        plannedOutputQty:
          type: number
        actualOutputQty:
          type: ["number", "null"]
        wastageQty:
          type: ["number", "null"]
        wastagePercent:
          type: ["number", "null"]
        yieldPercent:
          type: ["number", "null"]
        productionDate:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        responsibleUserId:
          type: ["integer", "null"]
        responsibleUserName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        inputMaterials:
          type: array
          items:
            $ref: "#/components/schemas/ProductionInput"
        outputProducts:
          type: array
          items:
            $ref: "#/components/schemas/ProductionInput"
        createdAt:
          type: string
    ProductionBatchList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ProductionBatch"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ProductionBatchInput:
      type: object
      required: [type, stageFromStoreId, stageToStoreId, plannedOutputQty, inputMaterials]
      properties:
        type:
          type: string
        stageFromStoreId:
          type: integer
        stageToStoreId:
          type: integer
        plannedOutputQty:
          type: number
        productionDate:
          type: ["string", "null"]
        responsibleUserId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        inputMaterials:
          type: array
          items:
            $ref: "#/components/schemas/ProductionInput"
    CompleteBatchInput:
      type: object
      required: [actualOutputQty, wastageQty, outputProducts]
      properties:
        actualOutputQty:
          type: number
        wastageQty:
          type: number
        notes:
          type: ["string", "null"]
        outputProducts:
          type: array
          items:
            $ref: "#/components/schemas/ProductionInput"

    # SALE
...
    Payment:
      type: object
      required: [id, saleId, amount, paymentDate, paymentMethod, createdAt]
      properties:
        id:
          type: integer
        saleId:
          type: integer
        invoiceNumber:
          type: ["string", "null"]
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        amount:
          type: number
        paymentDate:
          type: string
        paymentMethod:
          type: string
          enum: [cash, bank_transfer, cheque]
        bankName:
          type: ["string", "null"]
        reference:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    PaymentInput:
      type: object
      required: [saleId, amount, paymentDate, paymentMethod]
      properties:
        saleId:
          type: integer
        amount:
          type: number
        paymentDate:
          type: string
        paymentMethod:
          type: string
        bankName:
          type: ["string", "null"]
        reference:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # INVENTORY
    InventoryItem:
      type: object
      required: [id, productId, productName, storeId, storeName, quantity, unit]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productType:
          type: string
        storeId:
          type: integer
        storeName:
          type: string
        quantity:
          type: number
        unit:
          type: string
        reorderLevel:
          type: number
        isLowStock:
          type: boolean
    InventoryList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/InventoryItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    InventoryMovement:
      type: object
      required: [id, productId, storeId, movementType, quantity, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: ["string", "null"]
        storeId:
          type: integer
        storeName:
          type: ["string", "null"]
        movementType:
          type: string
          enum: [grn_receipt, transfer_out, transfer_in, production_input, production_output, sale, adjustment, wastage]
        quantity:
          type: number
        referenceId:
          type: ["integer", "null"]
        referenceType:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdBy:
          type: ["integer", "null"]
        createdAt:
          type: string
    InventoryMovementList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/InventoryMovement"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    InventoryAdjustmentInput:
      type: object
      required: [productId, storeId, quantity, reason]
      properties:
        productId:
          type: integer
        storeId:
          type: integer
        quantity:
          type: number
        reason:
          type: string
        notes:
          type: ["string", "null"]

    # DASHBOARD
...
    Notification:
      type: object
      required: [id, type, title, message, isRead, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        entityType:
          type: ["string", "null"]
        entityId:
          type: ["integer", "null"]
        createdAt:
          type: string
    NotificationList:
      type: object
      required: [data, total, unreadCount]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        total:
          type: integer
        unreadCount:
          type: integer

    # AUDIT
...